// src/_shared/enums/leonardo-styles.enum.ts
export enum ELeonardoPresetStyle {
  Anime = "ANIME",
  Bokeh = "BOKEH",
  Cinematic = "CINEMATIC",
  CinematicCloseup = "CINEMATIC_CLOSEUP",
  Creative = "CREATIVE",
  Dynamic = "DYNAMIC",
  Environment = "ENVIRONMENT",
  Fashion = "FASHION",
  Film = "FILM",
  Food = "FOOD",
  General = "GENERAL",
  HDR = "HDR",
  Illustration = "ILLUSTRATION",
  Leonardo = "LEONARDO",
  LongExposure = "LONG_EXPOSURE",
  Macro = "MACRO",
  Minimalistic = "MINIMALISTIC",
  Monochrome = "MONOCHROME",
  Moody = "MOODY",
  None = "NONE",
  Neutral = "NEUTRAL",
  Photography = "PHOTOGRAPHY",
  Portrait = "PORTRAIT",
  Raytraced = "RAYTRACED",
  Render3D = "RENDER_3D",
  Retro = "RETRO",
  SketchBw = "SKETCH_BW",
  SketchColor = "SKETCH_COLOR",
  StockPhoto = "STOCK_PHOTO",
  Vibrant = "VIBRANT",
  Unprocessed = "UNPROCESSED",
}

export enum ELeonardoPresetStyleId {
  Render3D = "debdf72a-91a4-467b-bf61-cc02bdeb69c6",
  Bokeh = "9fdc5e8c-4d13-49b4-9ce6-5a74cbb19177",
  Cinematic = "a5632c7c-ddbb-4e2f-ba34-8456ab3ac436",
  CinematicConcept = "33abbb99-03b9-4dd7-9761-ee98650b2c88",
  Creative = "6fedbf1f-4a17-45ec-84fb-92fe524a29ef",
  Dynamic = "111dc692-d470-4eec-b791-3475abac4c46",
  Fashion = "594c4a08-a522-4e0e-b7ff-e4dac4b6b622",
  GraphicDesignPopArt = "2e74ec31-f3a4-4825-b08b-2894f6d13941",
  GraphicDesignVector = "1fbb6a68-9319-44d2-8d56-2957ca0ece6a",
  HDR = "97c20e5c-1af6-4d42-b227-54d03d8f0727",
  Illustration = "645e4195-f63d-4715-a3f2-3fb1e6eb8c70",
  Macro = "30c1d34f-e3a9-479a-b56f-c018bbc9c02a",
  Minimalist = "cadc8cd6-7838-4c99-b645-df76be8ba8d8",
  Moody = "621e1c9a-6319-4bee-a12d-ae40659162fa",
  None = "556c1ee5-ec38-42e8-955a-1e82dad0ffa1",
  Portrait = "8e2bc543-6ee2-45f9-bcd9-594b6ce84dcd",
  PortraitFashion = "0d34f8e1-46d4-428f-8ddd-4b11811fa7c9",
  RayTraced = "b504f83c-3326-4947-82e1-7fe9e839ec0f",
  SketchBW = "be8c6b58-739c-4d44-b9c1-b032ed308b61",
  SketchColor = "093accc3-7633-4ffd-82da-d34000dfc0d6",
  StockPhoto = "5bdc3f2a-1be6-4d1c-8e77-992a30824a2c",
  Vibrant = "dee282d3-891f-4f73-ba02-7f8131e5541b",
}